Fixing build script and adding zoom levels.
Fixing build script so that it will upload a jar as well as a .aar file
to a maven repository.
Also adding zoom levels. This will allow for doing different things per
zoom level. One possible implementation is to change the image once a
certain zoom level is reached to provide a more detailed image. When
doing this, use `setImageDrawable(drawable, boolean)` with the second
argument set to false. This will prevent a flicker from appearing. When
using this method, however, the images that are being "hot-swapped"
should be the **exact same size**.